Created and linked pizza page to index
I created and linked the pizza page to index so it could be viewed by clicking on the word 'pizza'.
The website contains an image of pizza sources from the images sub-directory within the odin-recipes directory.
The pizza webpage file itself is located in its own recipes sub-directory along with 2 more files awaiting updates.